% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_results.R
\name{analyze_results}
\alias{analyze_results}
\title{Analyze Results}
\usage{
analyze_results(publications, before_placeholder_percentage)
}
\arguments{
\item{publications}{A data frame containing processed publication data.}

\item{before_placeholder_percentage}{The percentage of names alphabetically before the target name.}
}
\value{
A list of results, including statistical summaries and tests.
}
\description{
This function summarizes the results of the authorship analysis, comparing observed vs. theoretical shares,
first authorship rates, and performing a binomial test.
}
