% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_last_name.R
\name{extract_last_name}
\alias{extract_last_name}
\title{Extract Last Names}
\usage{
extract_last_name(author)
}
\arguments{
\item{author}{A string representing the full name of an author.}
}
\value{
The last name of the author.
}
\description{
A helper function to extract the last name from a full author name.
}
