% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_publications.R
\name{fetch_publications}
\alias{fetch_publications}
\title{Fetch Publications from Google Scholar}
\usage{
fetch_publications(profile_id)
}
\arguments{
\item{profile_id}{The Google Scholar profile ID.}
}
\value{
A data frame of publications.
}
\description{
This function retrieves publications using the \code{scholar} package.
}
