% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_journals.R
\name{filter_journals}
\alias{filter_journals}
\title{Filter Journals by Whitelist}
\usage{
filter_journals(publications)
}
\arguments{
\item{publications}{A data frame of publications.}
}
\value{
A filtered data frame of publications.
}
\description{
This function filters publications to include only those from a predefined whitelist.
}
