% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_authorship.R
\name{process_authorship}
\alias{process_authorship}
\title{Process Authorship Data}
\usage{
process_authorship(publications, last_name, before_placeholder_percentage)
}
\arguments{
\item{publications}{A data frame containing publication data.}

\item{last_name}{The last name of the author being analyzed (e.g., "Broekel").}

\item{before_placeholder_percentage}{The percentage of surnames alphabetically before the author's last name.}
}
\value{
A data frame with additional columns for authorship analysis.
}
\description{
This function processes authorship information for each publication, calculating co-author statistics
and probabilities of first authorship based on alphabetical order.
}
