% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_surname_distribution.R
\name{read_surname_distribution}
\alias{read_surname_distribution}
\title{Read and Process Surname Distribution}
\usage{
read_surname_distribution(file_name, placeholder_name)
}
\arguments{
\item{file_name}{The name of the Excel file containing surname data (should be in the data directory).}

\item{placeholder_name}{The name to use for alphabetical comparison.}
}
\value{
A list containing the distribution data and the percentage of names before the placeholder.
}
\description{
This function reads the surname distribution data from an Excel file and calculates
the percentage of names alphabetically before a specified name.
}
