% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_journals.R
\name{standardize_journals}
\alias{standardize_journals}
\title{Standardize Journal Names}
\usage{
standardize_journals(publications)
}
\arguments{
\item{publications}{A data frame of publications.}
}
\value{
A data frame with standardized journal names.
}
\description{
This function standardizes journal names using a mapping file.
}
