% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_results.R
\name{analyze_results}
\alias{analyze_results}
\title{Analyze Results}
\usage{
analyze_results(publications)
}
\arguments{
\item{publications}{A data frame of processed publications.}
}
\value{
A list of results including observed vs theoretical shares, first authorship analysis, and binomial test.
}
\description{
This function calculates summary statistics and performs statistical tests on the processed publications data.
}
